tests: fix a test on 32-bit builds
We were trying to squeeze 64-bit test vectors into gsize, which is fine
on 64bit systems but doesn't work very well on 32-bit.

Change that to a guint64.

+++ glib/tests/overflow.c
@@ -16,7 +16,7 @@
 typedef struct
 {
   gboolean success;
-  gsize c, a, b;
+  guint64 c, a, b;
 } Case;
 
 static void
